backend/requirements.txt
pydantic>=2.0.0
sqlalchemy>=2.0.0
pytest>=8.0.0
pdfplumber>=0.10.0
pymupdf>=1.24.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
psycopg[binary]>=3.2.1
